finished the email reservations for dropped reservations
