+++ app-backend/api/email-system/emailSystem.py
@@ -1,38 +1,35 @@
-from api.services import users
+from .. services import users
 import smtplib
+from sqlmodel import Session
 import ssl
 from dotenv import load_dotenv
 from email.message import EmailMessage
 import os
 load_dotenv()
 
-def EmailDroppedReservation(reason,id):
-    userinfo = users.fetch_users_by_id(id)
-
-    print(userinfo)
+def EmailDroppedReservation(id, reason):
+    userInfo = users.fetch_users_by_id(id)
+    print(userInfo)
     subject = "Computer reservation has been canceled"
-    body = "Your reservation for" + "date" + " at " + "time" + "has been canceled for" + reason
-    receiver_email = userinfo.email
-    SendEmail(subject, body, receiver_email)
+    body = "Your reservation has been canceled for" + reason + " "
+    receiverEmail = userInfo.email
+    SendEmail(subject, body, receiverEmail)
 
-
-def SendEmail(subject, body, receiver_email):
-    sender_email ="csc4990librarysmtp@gmail.com"
-    password = os.getenv('PASSWORD')
+def SendEmail(subject, body, receiverEmail):
+    senderEmail = os.getenv('BACKENDEMAIL')
+    password = os.getenv('BACKENDEMAILPASSWORD')
     message = EmailMessage()
     message.set_content(body)
     message['Subject'] = subject
-    message['From'] = sender_email
-    message['To'] = receiver_email
+    message['From'] = senderEmail
+    message['To'] = receiverEmail
 
     context = ssl.create_default_context()
 
     try:
         with smtplib.SMTP_SSL("smtp.gmail.com", 465, context=context) as server:
-            server.login(sender_email, password)
+            server.login(senderEmail, password)
             server.send_message(message)
             print("Email sent successfully!")
     except smtplib.SMTPException as e:
         print(f"Error: {e}")
-
-EmailDroppedReservation()
